refactor(mobile): split channel detail widgets
